[ Corrected program workflow and added secret handling from .env file ]
