+++ utils/llm_groq.py
@@ -1,5 +1,8 @@
 import os
 from groq import Groq
+from dotenv import load_dotenv
+
+load_dotenv(os.path.join(os.path.dirname(__file__), '..', '.env'))
 
 groq_client = Groq(api_key=os.getenv("GROQ_API_KEY"))
