+++ utils/vectorized_db.py
@@ -1,8 +1,11 @@
 import os
 import ollama
+from dotenv import load_dotenv
 from pinecone import Pinecone
 from crawl4ai.models import CrawlResult
 from langchain.text_splitter import RecursiveCharacterTextSplitter
+
+load_dotenv(os.path.join(os.path.dirname(__file__), '..', '.env'))
 
 pc = Pinecone(api_key=os.getenv("PINECONE_API_KEY"))
 index = pc.Index(os.getenv("PINECONE_INDEX_NAME"))
@@ -10,12 +13,28 @@ index = pc.Index(os.getenv("PINECONE_INDEX_NAME"))
 def generate_embedding_from_text(text: str) -> list[float]:
     response = ollama.embed(model="nomic-embed-text", input=text)
     embedding = response["embeddings"]
+
+    if embedding[0] and isinstance(embedding[0], list):
+        embedding = [item for sublist in embedding for item in sublist]
     return embedding
+
+def normalize_urls(url):
+    normalized_url = (
+        url.replace("https://", "")
+        .replace("www.", "")
+        .replace("/", "_")
+        .replace("-", "_")
+        .replace(".", "_")
+    )
+
+    return normalized_url
 
 def add_text_to_index(results: list[CrawlResult]):
     for result in results:
-        # Extract text content from markdown
-        text_content = result.markdown_content
+        if result is None or result.markdown_v2 is None:
+            continue
+
+        text_content = result.markdown_v2.fit_markdown
         
         # Split text into chunks
         text_splitter = RecursiveCharacterTextSplitter(
@@ -32,23 +51,24 @@ def add_text_to_index(results: list[CrawlResult]):
         ids = []
         
         for i, chunk in enumerate(chunks):
-            # Generate embedding for this chunk
             embedding = generate_embedding_from_text(chunk)
+
+            normalized_url = normalize_urls(result.url)
+            chunk_id = f"{normalized_url}_{i}"
+
             
-            # Create unique ID for this chunk
-            chunk_id = f"{result.url}_{i}"
-            
-            # Prepare metadata
             metadata = {
-                "url": result.url,
-                "title": result.title,
-                "timestamp": result.timestamp,
+                "url": normalized_url,
+                # "title": result.title,
+                # "timestamp": result.timestamp,
                 "chunk_index": i,
                 "total_chunks": len(chunks),
                 "text": chunk
             }
-            
-            # Add to lists
+
+            if len(chunk) <= 3: # there are times when the chunk is too small like 2 words
+                continue
+
             documents.append(chunk)
             embeddings.append(embedding)
             metadatas.append(metadata)
@@ -75,7 +95,7 @@ def retrieve_context(query: str) -> list[dict[str, str]]:
     )
 
     results = []
-    for match in response:
+    for match in response.matches:
         metadata = match.metadata
 
         results.append({"url": metadata["url"], "text": metadata["text"]})
